change package structure, move SpringBootConsoleApplication.java to com.parkingshare.console
