+++ shared-parking-spot-app/src/main/java/com/parkingsharing/sql/Company.java
@@ -1,5 +1,5 @@
 package com.parkingsharing.sql;
-// Generated Mar 23, 2024, 7:54:27 PM by Hibernate Tools 6.4.4.Final
+// Generated 26 mar 2024, 16:32:33 by Hibernate Tools 6.4.4.Final
 
 
 import jakarta.persistence.Column;
